#1756-Advanced-DB-entry-for-workflow
- Complete overhaul of module structure for workload, workload-dynamic and workload-static
- Using of registry and extension points for workload static and dynamic
