+++ inception-app-webapp/pom.xml
@@ -293,10 +293,6 @@
       <groupId>de.tudarmstadt.ukp.clarin.webanno</groupId>
       <artifactId>webanno-doc</artifactId>
     </dependency>
-    <dependency>
-      <groupId>de.tudarmstadt.ukp.inception.app</groupId>
-      <artifactId>inception-workload-dynamic</artifactId>
-    </dependency>
 
     <!-- UIMA DEPENDENCIES -->
     <dependency>
@@ -519,7 +515,11 @@
 	</dependency>
       <dependency>
           <groupId>de.tudarmstadt.ukp.inception.app</groupId>
-          <artifactId>inception-workload-dynamic</artifactId>
+          <artifactId>inception-workload</artifactId>
+      </dependency>
+      <dependency>
+          <groupId>de.tudarmstadt.ukp.inception.app</groupId>
+          <artifactId>inception-workload-static</artifactId>
       </dependency>
   </dependencies>
   <build>
