+++ inception-app-webapp/src/main/resources/db/changelog/db.changelog-master.xml
@@ -34,5 +34,5 @@
   <include file="de/tudarmstadt/ukp/inception/recommendation/imls/stringmatch/model/db-changelog.xml"/>
   <include file="de/tudarmstadt/ukp/inception/recommendation/imls/weblicht/model/db-changelog.xml"/>
   <include file="de/tudarmstadt/ukp/inception/curation/db-changelog.xml"/>
-  <include file="de/tudarmstadt/ukp/inception/workload/dynamic/model/db-changelog.xml"/>
+  <include file="de/tudarmstadt/ukp/inception/workload/model/db-changelog.xml"/>
 </databaseChangeLog>
