--- inception-workload-dynamic/src/main/java/de/tudarmstadt/ukp/inception/workload/dynamic/api/WorkloadConst.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2020
- * Ubiquitous Knowledge Processing (UKP) Lab
- * Technische Universität Darmstadt
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.tudarmstadt.ukp.inception.workload.dynamic.api;
-
-public class WorkloadConst
-{
-    //Workflow
-    public static final String DEFAULT_WORKFLOW = "Default workflow";
-    public static final String DYNAMIC_WORKFLOW = "Dynamic workflow";
-
-    //Workload
-    public static final String DEFAULT_MONITORING = "Default monitoring";
-    public static final String WORKLOAD_MONITORING = "Workload monitoring";
-
-    //Workflow document distribution type
-    public static final String DEFAULT_WORKFLOW_TYPE = "Default";
-    public static final String RANDOMIZED_WORKFLOW_TYPE = "Randomized";
-}
