--- inception-workload-dynamic/src/main/java/de/tudarmstadt/ukp/inception/workload/dynamic/extensions/DynamicDocumentNavigator.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2020
- * Ubiquitous Knowledge Processing (UKP) Lab
- * Technische Universität Darmstadt
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package de.tudarmstadt.ukp.inception.workload.dynamic.extensions;
-
-import org.apache.wicket.markup.html.panel.Panel;
-
-public class DynamicDocumentNavigator extends Panel
-{
-    public DynamicDocumentNavigator(String aID)
-    {
-        super(aID);
-    }
-}
