--- inception-workload-dynamic/src/main/java/de/tudarmstadt/ukp/inception/workload/dynamic/page/settings/WorkflowAndMonitoringPanel.properties
@@ -1,27 +1,0 @@
-# Copyright 2020
-# Ubiquitous Knowledge Processing (UKP) Lab
-# Technische Universität Darmstadt
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-panel=panel
-workflow=Please select a workflow manager
-defaultWorkflow=Default workflow manager
-
-monitoring=Please select a monitoring page
-defaultMonitoring=Default monitoring page
-
-#Card Header
-Monitoring=Monitoring
-Workflow=Workflow
-Settings=Settings
