+++ inception-workload-dynamic/src/main/java/de/tudarmstadt/ukp/inception/workload/dynamic/workflow/DynamicWorkflowDocumentNavigationActionBarExtension.html
