+++ inception-workload-dynamic/src/main/java/de/tudarmstadt/ukp/inception/workload/dynamic/workload/WorkloadPageMenuItem.java
@@ -1,27 +1,26 @@
 /*
- * Copyright 2018
+ * Copyright 2020
  * Ubiquitous Knowledge Processing (UKP) Lab
  * Technische Universität Darmstadt
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *  
+ *
  *  http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.inception.app.menu;
+package de.tudarmstadt.ukp.inception.workload.dynamic.workload;
 
 import org.apache.wicket.Page;
 import org.apache.wicket.Session;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.ProjectService;
 import de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst;
@@ -29,14 +28,12 @@ import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 import de.tudarmstadt.ukp.clarin.webanno.security.UserDao;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.User;
 import de.tudarmstadt.ukp.clarin.webanno.ui.core.menu.MenuItem;
-import de.tudarmstadt.ukp.clarin.webanno.ui.monitoring.page.MonitoringPage;
 import de.tudarmstadt.ukp.inception.ui.core.session.SessionMetaData;
-import de.tudarmstadt.ukp.inception.workload.monitoring.extension.StaticWorkloadExtension;
+import de.tudarmstadt.ukp.inception.workload.dynamic.extension.DynamicWorkloadExtension;
 import de.tudarmstadt.ukp.inception.workload.registry.WorkloadRegistry;
 
-@Component
-@Order(300)
-public class MonitoringPageMenuItem implements MenuItem
+@Order(500)
+public class WorkloadPageMenuItem implements MenuItem
 {
     private @Autowired UserDao userRepo;
     private @Autowired ProjectService projectService;
@@ -45,21 +42,21 @@ public class MonitoringPageMenuItem implements MenuItem
     @Override
     public String getPath()
     {
-        return "/monitoring";
+        return "/workload";
     }
-    
+
     @Override
     public String getIcon()
     {
         return "images/attribution.png";
     }
-    
+
     @Override
     public String getLabel()
     {
-        return "Monitoring";
+        return "Workload";
     }
-    
+
     /**
      * Only admins and project managers can see this page
      */
@@ -70,22 +67,25 @@ public class MonitoringPageMenuItem implements MenuItem
         if (sessionProject == null) {
             return false;
         }
+
         // The project object stored in the session is detached from the persistence context and
         // cannot be used immediately in DB interactions. Fetch a fresh copy from the DB.
 
         // Visible if the current user is a curator or project admin
         User user = userRepo.getCurrentUser();
+
         return (projectService.isCurator(sessionProject, user)
-                || projectService.isProjectAdmin(sessionProject, user))
-                && WebAnnoConst.PROJECT_TYPE_ANNOTATION.equals(sessionProject.getMode())
-                && workloadRegistry.getSelectedExtension(sessionProject).
-                equals(StaticWorkloadExtension.class.toString().
-                substring(StaticWorkloadExtension.class.toString().lastIndexOf(".") + 1));
+            || projectService.isProjectAdmin(sessionProject, user))
+            && WebAnnoConst.PROJECT_TYPE_ANNOTATION.equals(sessionProject.getMode())
+            && workloadRegistry.getSelectedExtension(sessionProject).
+            equals(DynamicWorkloadExtension.class.toString().
+                substring(DynamicWorkloadExtension.class.toString().lastIndexOf(".") + 1));
+
     }
-    
+
     @Override
     public Class<? extends Page> getPageClass()
     {
-        return MonitoringPage.class;
+        return DynamicWorkloadManagementPage.class;
     }
 }
