--- inception-workload-dynamic/src/main/resources/de/tudarmstadt/ukp/inception/workload/dynamic/model/db-changelog.xml
@@ -1,36 +1,0 @@
-<?xml version="1.1" encoding="UTF-8" standalone="no"?>
-<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.5.xsd">
-  <changeSet author="Christoph Scheunemann" id="20200408-01">
-    <createTable tableName="workload_assignment">
-      <column autoIncrement="true" name="id" type="BIGINT">
-        <constraints primaryKey="true" />
-      </column>
-      <column name="project" type="BIGINT">
-        <constraints nullable="false" />
-      </column>
-      <column name="workflowmanagertype" type="VARCHAR(50)">
-        <constraints nullable="false" />
-      </column>
-        <column name="workloadmanagertype" type="VARCHAR(50)">
-            <constraints nullable="false" />
-        </column>
-        <column name="workflow" type="VARCHAR(50)">
-            <constraints nullable="false" />
-        </column>
-        <column name="numberofannotations" type="int">
-            <constraints nullable="false" />
-        </column>
-        <column name="traits" type="CLOB(64000)"></column>
-    </createTable>
-    <createIndex indexName="FK_project_id" tableName="workload_assignment">
-      <column name="project" />
-    </createIndex>
-    <addForeignKeyConstraint constraintName="FK_project_id"
-      baseTableName="workload_assignment" baseColumnNames="project"
-      referencedTableName="project" referencedColumnNames="id"
-      deferrable="false" initiallyDeferred="false" 
-      onDelete="CASCADE" onUpdate="CASCADE"/>
-  </changeSet>
-</databaseChangeLog>
