+++ inception-workload/src/main/java/de/tudarmstadt/ukp/inception/workload/settings/WorkloadPanel.html
@@ -1,0 +1,66 @@
+<!DOCTYPE html>
+<!--
+ - Copyright 2020
+ - Ubiquitous Knowledge Processing (UKP) Lab
+ - Technische Universität Darmstadt
+ -
+ - Licensed under the Apache License, Version 2.0 (the "License");
+ - you may not use this file except in compliance with the License.
+ - You may obtain a copy of the License at
+ -
+ - http://www.apache.org/licenses/LICENSE-2.0
+ -
+ - Unless required by applicable law or agreed to in writing, software
+ - distributed under the License is distributed on an "AS IS" BASIS,
+ - WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ - See the License for the specific language governing permissions and
+ - limitations under the License.
+-->
+<html xmlns:wicket="http://wicket.apache.org">
+<head>
+  <wicket:head>
+    <style type="text/css">
+      input[type="radio"], label {
+
+        vertical-align: middle;
+        padding: 4px;
+        margin: 4px;
+      }
+    </style>
+  </wicket:head>
+</head>
+
+<body>
+<wicket:panel>
+  <div class="scrolling flex-content flex-h-container flex-gutter">
+    <form wicket:id="form" class="col-sm-5 flex-gutter">
+      <div class="card m-2 ">
+        <div class="card-header">
+          <wicket:message key="Settings"/>
+        </div>
+        <div class="card-body">
+        <div class="card m-2 ">
+          <div class="card-header">
+            <wicket:message key="Monitoring"/>
+          </div>
+         <div class="card-body">
+           <label class="col-form-label" >
+             <wicket:message key="monitoring"/>
+           </label>
+         <br />
+         <div>
+           <span wicket:id="monitoringRadios"></span>
+         </div>
+       </div>
+     </div>
+     <button class="btn btn-primary" wicket:id="save">
+       <i class="fas fa-save"></i>&nbsp;
+       <wicket:message key="save"/>
+     </button>
+    </div>
+   </div>
+  </form>
+ </div>
+</wicket:panel>
+</body>
+</html>
