+++ pom.xml
@@ -97,7 +97,6 @@
     <module>inception-testing</module>
     <module>inception-ui-core</module>
     <module>inception-app-webapp</module>
-    <module>inception-workload-dynamic</module>
 
     <!-- Internal search -->
     <module>inception-ui-search</module>
@@ -134,6 +133,9 @@
     <module>inception-doc</module>
     <!-- Curation -->
     <module>inception-curation</module>
+    <module>inception-workload</module>
+    <module>inception-workload-static</module>
+    <module>inception-workload-dynamic</module>
   </modules>
   <repositories>
     <!-- For RELEASEes of WebAnno / DKPro Core-->
@@ -283,6 +285,21 @@
       </dependency>
       <dependency>
         <groupId>de.tudarmstadt.ukp.inception.app</groupId>
+        <artifactId>inception-workload</artifactId>
+        <version>0.17.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>de.tudarmstadt.ukp.inception.app</groupId>
+        <artifactId>inception-workload-static</artifactId>
+        <version>0.17.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>de.tudarmstadt.ukp.inception.app</groupId>
+        <artifactId>inception-workload-dynamic</artifactId>
+        <version>0.17.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>de.tudarmstadt.ukp.inception.app</groupId>
         <artifactId>inception-external-search-core</artifactId>
         <version>0.17.0-SNAPSHOT</version>
       </dependency>
@@ -419,11 +436,6 @@
       <dependency>
         <groupId>de.tudarmstadt.ukp.inception.app</groupId>
         <artifactId>inception-search-mtas</artifactId>
-        <version>0.17.0-SNAPSHOT</version>
-      </dependency>
-      <dependency>
-        <groupId>de.tudarmstadt.ukp.inception.app</groupId>
-        <artifactId>inception-workload-dynamic</artifactId>
         <version>0.17.0-SNAPSHOT</version>
       </dependency>
